Revert "rtp: Release the request pads from gstrtpbin, require gst-p-bad 0.10.13 to match"
This reverts commit f5c228f61018b75d908e1c1ca586fdb00a587500.

The session dispose then tries to remove the recv part of the gstrtpbin, which
tries to stop the jitterbuffer stream. The session dispose could be called
from a streaming thread (if it is disposed during a src-pad-added signal for example). We need to stop it from another thread in this case.