rawconference: Remove stream from session in stream's dispose.
There's a chance that removing the stream when the session has it
weak-reffed can be called from a streaming thread. This can cause
it to crash and/or deadlock. This patch changes the stream to
call the remove_stream function in the session in its dispose
function. The stream already protects itself from being disposed
in a streaming thread and therefore prevents the crash/deadlock.