rawudp: Send the original STUN packets from the same thread
This is to make sure the STUN packets for each component are sent in order.
Some firewalls just allocate output ports linearly. In that case, we prefer
if the order of the outside ports is the same as the order of the local ports,
that way, if we do UPnP, we have fewer chances of mixing it up.