rtpcodecspecific: Check if config params have changed and need to be re-sent
Check if matching codecs have different config parameters, and return the list
of codecs that need to be resent or NULL otherwise.
We do that by matching the codecs one by one
with a fuzzy SDP Offer/Answer match and then when we find matchs, we check
if any of the config params are either different or only available in the new
or the old one.