shm-stream-transmitter: Add property to control the buffer-time
Allow the transmitter to buffer how much can be maximally queued in the
shmsink before blocking. Default to 20 miliseconds.

For audio network packets tend to be around 20ms is size, so 20ms is a
suitable size for audio.

For video, 20 miliseconds is less then one frame at 30 fps, so with 20ms
as default at most one video frame will be queued up in the shm sink.